[extension/oidcauth] Add static public_keys_file to configuration
When provided, the extension will attempt to load a static set of public
keys from a JWKS file at the given path. These keys (and only these
keys) will be used to verify the signatures of incoming tokens.
Discovery will be disabled for any provider so configured.

In addition, the extension will set up a file watcher for each directory
referenced in this way, and will hot-reload the static keys in the JWKS
file whenever a change is noticed. This allows keys to be rotated
without necessarily needing to restart the collector. We extend the
`providerContainer` with a `Verify` method, so that we can encapsulate
the decision of "should we grab the mutex" in one location (we only need
to grab the mutex when using static configuration - it's unused
otherwise and we can avoid paying the cost of grabbing the lock).

Due to limitations of the underlying `go-oidc` library, only RSA, ECDSA,
and ED25519 public keys are supported.